one chuser() to rule them all
